Add dto files && add logger.interceptor

+++ src/products/products.controller.ts
@@ -1,6 +1,17 @@
-import { Body, Controller, Delete, Get, Param, Post } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Put,
+  Delete,
+  Param,
+  Body,
+  ParseUUIDPipe,
+  NotFoundException,
+} from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -12,18 +23,35 @@ export class ProductsController {
   }
 
   @Get('/:id')
-  getById(@Param('id') id: string) {
-    return this.productsService.getById(id);
-  }
-
-  @Delete('/:id')
-  deleteById(@Param('id') id: string) {
-    this.productsService.deleteById(id);
-    return { success: true };
+  getById(@Param('id', new ParseUUIDPipe()) id: string) {
+    const product = this.productsService.getById(id);
+    if (!product) throw new NotFoundException('Product not found');
+    return product;
   }
 
   @Post('/')
   create(@Body() productData: CreateProductDTO) {
     return this.productsService.create(productData);
+  }
+
+  @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
+
+  @Delete('/:id')
+  deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+    if (!this.productsService.getById(id)) {
+      throw new NotFoundException('Product not found');
+    }
+    this.productsService.deleteById(id);
+    return { success: true };
   }
 }
